+++ demo/Enqueing/MyJobClass.php
@@ -14,6 +14,6 @@ class MyJobClass
     public function perform(Job $job): void
     {
         // ...
-        echo 'Perform ', $job->getId(), ' job', PHP_EOL;
+        echo 'Perform ', $job->jid, ' job', PHP_EOL;
     }
 }